test-with-pom - move register-test-user into pom/sequences
