+++ e2e/user-can-cancel-add-dog.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "@playwright/test";
-import { registerTestUser } from "./_lib/register-test-user";
+import { registerTestUser } from "./_lib/pom/sequences/register-test-user";
 import { getTestBirthday } from "./_lib/e2e-test-utils";
 import { generateTestDogName } from "./_lib/e2e-test-utils";
 import { urlOf } from "./_lib/e2e-test-utils";
